Added check to make sure trackday has no members prior to deletion

+++ api/controllers/trackdayController.js
@@ -702,6 +702,10 @@ exports.trackday_delete = [
 
     asyncHandler(async (req, res, next) => {
         if (req.user.memberType === 'admin') {
+            // Make sure trackday has no members
+            const trackday = await Trackday.findById(req.params.trackdayID).exec();
+            if (trackday.members.length) return res.status(400).send({ msg: 'Trackday has registered users' })
+
             await Trackday.findByIdAndDelete(req.params.trackdayID);
             logger.info({ message: "trackday deleted: " + req.params.trackdayID });
             return res.sendStatus(200);
